TST: move fixture to pandas/conftest.py
