+++ pandas/conftest.py
@@ -437,6 +437,26 @@ def index(request):
 index_fixture2 = index
 
 
+@pytest.fixture(params=indices_dict.keys())
+def index_with_missing(request):
+    """
+    Fixture for indices with missing values
+    """
+    if request.param in ["datetime", "datetime-tz", "period", "timedelta"]:
+        pytest.xfail("stable descending order sort not implemented")
+    if request.param in ["int", "uint", "range", "empty", "repeats"]:
+        pytest.xfail("missing values not supported")
+    if request.param in ["categorical", "tuples", "mi-with-dt64tz-level", "multi"]:
+        pytest.xfail("missing value sorting order not defined for index type")
+    ind = indices_dict[request.param].copy()
+    vals = ind.values
+    print(request.param)
+    vals[0] = None
+    vals[-1] = None
+    ind = type(ind)(vals)
+    return ind
+
+
 # ----------------------------------------------------------------
 # Series'
 # ----------------------------------------------------------------
